+++ madsim/src/sim/task/mod.rs
@@ -22,7 +22,7 @@ use std::{
         Arc, Weak,
     },
     task::{Context, Poll, Waker},
-    time::Duration,
+    time::{Duration, Instant},
 };
 use tokio::sync::watch;
 use tracing::{debug, error, error_span, trace, Span};
@@ -75,6 +75,9 @@ pub struct TaskInfo {
     pub span: Span,
     /// The spawn location.
     location: StaticLocation,
+    /// The time when this task is spawned.
+    #[allow(dead_code)]
+    spawn_time: Instant,
     /// The waker of this task. Used for cancellation.
     waker: Waker,
     /// A flag indicating that the task has been cancelled.
@@ -117,6 +120,7 @@ impl NodeInfo {
             name,
             node: self.clone(),
             location: Location::caller(),
+            spawn_time: Instant::now(),
             waker: futures_util::task::noop_waker(), // updated later
             cancelled: AtomicBool::new(false),
         });
@@ -137,6 +141,20 @@ impl NodeInfo {
         let mut tasks = self.tasks.lock();
         tasks.retain(|weak| weak.strong_count() != 0);
         tasks.len()
+    }
+
+    fn num_tasks_by_spawn(&self) -> BTreeMap<String, usize> {
+        let mut tasks = self.tasks.lock();
+        let mut map = BTreeMap::new();
+        tasks.retain(|weak| {
+            if let Some(task) = weak.upgrade() {
+                *map.entry(task.location.to_string()).or_default() += 1;
+                true
+            } else {
+                false
+            }
+        });
+        map
     }
 
     pub(crate) fn is_killed(&self) -> bool {
@@ -468,6 +486,7 @@ impl TaskHandle {
             .map(|node| node.info.num_tasks())
             .sum()
     }
+
     pub fn num_tasks_by_node(&self) -> BTreeMap<String, usize> {
         self.nodes
             .lock()
@@ -482,6 +501,24 @@ impl TaskHandle {
                 )
             })
             .collect()
+    }
+
+    pub fn num_tasks_by_node_by_spawn(&self) -> String {
+        let map = self
+            .nodes
+            .lock()
+            .values()
+            .map(|node| {
+                (
+                    node.info
+                        .name
+                        .clone()
+                        .unwrap_or_else(|| format!("node-{}", node.info.id)),
+                    node.info.num_tasks_by_spawn(),
+                )
+            })
+            .collect::<BTreeMap<String, _>>();
+        format!("{map:#?}")
     }
 }